requirements.txt
claude-agent-sdk==0.1.52
langfuse>=2.0.0
google-cloud-firestore>=2.19.0
google-cloud-bigquery>=3.41.0
fastapi>=0.135.2
uvicorn>=0.42.0
pydantic>=2.12.5
pytest>=9.0.2
pytest-asyncio>=1.3.0
isort>=8.0.1
black>=26.3.1
ruff>=0.15.8
pre-commit>=4.5.1